feat(main): COM-2736 remove and replace components
- remove Form from main page
- change Industries origin from main folder to service since it has been removed from main page and now exists only on outsourcing-front-end page
- replace and refactor Card components, because now two pages share it
- fix header styles
