+++ components/main/Projects/Projects.jsx
@@ -6,7 +6,7 @@ import styles from './Projects.styles'
 import Heading from '../../ui-kit/core-design/Heading'
 import Grid from '../../ui-kit/core-design/Grid'
 import Post from './Post'
-import Card from '../Card'
+import Card from '../../Card'
 import posts from '../../../data/main/posts'
 import Solutions from '../../our-cases/Solutions'
 
@@ -28,7 +28,7 @@ const Projects = ({ className, l10n: { translations, language } }) => {
 
       {language === 'en' && <Solutions className="solutions" />}
 
-      {projects.map(({ id, title, description, href, images, imagesHovered }) => {
+      {projects.map(({ id, title, description, href, images, alt, imagesHovered }) => {
         const Player = () => (
           <div className="player-wrapper">
             <iframe
@@ -48,6 +48,7 @@ const Projects = ({ className, l10n: { translations, language } }) => {
           <Card
             testId={`Projects:${id === 'radio' ? 'block' : 'link'}.${id}`}
             className={cn('card', `card_${id}`)}
+            alt={alt}
             key={id}
             id={id}
             title={title}
