+++ data/main/projects.js
@@ -17,6 +17,7 @@ export default [
     id: 'tracker',
     title: (t) => t.main.projects.tracker.title,
     description: (t) => t.main.projects.tracker.description,
+    alt: (t) => t.main.imgAlt.tracker,
     href: 'https://tracker.csssr.com/',
     images: trackerImages,
     imagesHovered: trackerImagesHovered,
@@ -25,6 +26,7 @@ export default [
     id: 'school',
     title: (t) => t.main.projects.school.title,
     description: (t) => t.main.projects.school.description,
+    alt: (t) => t.main.imgAlt.school,
     href: 'https://school.csssr.com',
     images: schoolImages,
     imagesHovered: schoolImagesHovered,
@@ -33,6 +35,7 @@ export default [
     id: 'blog',
     title: (t) => t.main.projects.blog.title,
     description: (t) => t.main.projects.blog.description,
+    alt: (t) => t.main.imgAlt.blog,
     href: 'https://blog.csssr.com',
     images: blogImages,
     imagesHovered: blogImagesHovered,
@@ -41,6 +44,7 @@ export default [
     id: 'radio',
     title: (t) => t.main.projects.radio.title,
     description: (t) => t.main.projects.radio.description,
+    alt: (t) => t.main.imgAlt.radio,
     href: 'https://soundcloud.com/csssr',
     images: radioImages,
   },
