+++ static/locales/en/outsourcingFrontEnd.json
@@ -35,6 +35,26 @@
       }
     ]
   },
+  "industries": {
+    "title": "Industries",
+    "subtitle": "We&nbsp;take a&nbsp;goal-oriented approach across a&nbsp;wide range of&nbsp;industries",
+    "fintech": {
+      "title": "Fintech",
+      "description": "We&rsquo;ve built a&nbsp;strong digital presence for many banking and financial services."
+    },
+    "eCommerce": {
+      "title": "E-Commerce",
+      "description": "We&nbsp;help our clients in&nbsp;the highly competitive <span style=\"white-space: nowrap;\">e-commerce</span> industry win each new customer."
+    },
+    "information": {
+      "title": "Information Technology",
+      "description": "We&nbsp;push IT-companies to&nbsp;the cutting edge of&nbsp;technological progress."
+    },
+    "mediaAndMarketing": {
+      "title": "Media &amp;&nbsp;Marketing",
+      "description": "We&nbsp;deliver high-quality experiences capable of&nbsp;handling large-scale data and any scale of&nbsp;concurrent users."
+    }
+  },
   "whatWeDo": {
     "title": "What We&nbsp;Do",
     "list": [
@@ -109,6 +129,10 @@
     "ourAdvantages": [
       ["We work absolutely remotely so we have elaborated reliable remote processes."],
       ["We are focused on a narrow niche of front-end development.", "We can deal with both short- and long-term tasks."]
-    ]
+    ],
+    "fintech": "CSSSR builds a strong digital presence for most banking and financial services.",
+    "eCommerce": "CSSSR makes our clients in the highly competitive e-commerce industry win customers.",
+    "information": "We push IT-companies on the cutting edge of technological progress.",
+    "mediaAndMarketing": "Taking advantage of its profound high-quality experience, CSSSR is capable of handling large-scale data and any scale of concurrent users."
   }
 }
