+++ static/locales/ru/main.json
@@ -33,26 +33,6 @@
       "link": "Читать о технологиях"
     }
   },
-  "industries": {
-    "title": "Индустрии",
-    "subtitle": "Помогаем бизнесу в&nbsp;разработке веб-решений любой сложности",
-    "fintech": {
-      "title": "Финтех",
-      "description": "Обеспечиваем мощное цифровое присутствие для клиентов финансовой сферы."
-    },
-    "eCommerce": {
-      "title": "E-Commerce",
-      "description": "Боремся за&nbsp;каждого клиента в&nbsp;высококонкурентной индустрии интернет-торговли."
-    },
-    "information": {
-      "title": "Информационные технологии",
-      "description": "Выводим IT-компании на&nbsp;новый технологический уровень."
-    },
-    "mediaAndMarketing": {
-      "title": "Медиа и маркетинг",
-      "description": "Помогаем создавать сервисы, к&nbsp;которым хочется возвращаться."
-    }
-  },
   "aboutUs": {
     "title": "Мы действительно любим",
     "description": "CSSSR&nbsp;&mdash; группа компаний с&nbsp;головным офисом в&nbsp;Сингапуре. Мы&nbsp;все работаем удалённо и&nbsp;убеждены&nbsp;&mdash; действительно хорошие решения появляются только в&nbsp;атмосфере свободы и&nbsp;взаимоуважения.",
@@ -112,15 +92,7 @@
   },
   "imgAlt": {
     "hero": "CSSSR — международная IT группа, специализирующаяся на фронтенд разработке.",
-    "outsourcing": "",
-    "backend": "",
-    "mvp": "",
-    "design": "",
     "ourFeatures": "Человек, аккуратно и педантично подстригающий геометрический бонсай, символизирует наше стремление делать всё идеально",
-    "fintech": "CSSSR обеспечивает мощное цифровое присутствие для клиентов финансовой сферы.",
-    "eCommerce": "CSSSR борется за каждого клиента в высококонкурентной индустрии интернет-торговли.",
-    "information": "CSSSR — группа компаний, специализирующаяся на фронтенд аутсорсе",
-    "mediaAndMarketing": "CSSSR обеспечивает высокое качество работы с большим объемом данных.",
     "aboutUs": "Человек обнимающий сердце c надписью «Делаем»",
     "volleyball": "Команды CSSSR после турнира по волейболу",
     "elbrus": "Флаг CSSSR на Эльбрусе",
