+++ static/locales/ru/outsourcingFrontEnd.json
@@ -35,6 +35,26 @@
       }
     ]
   },
+  "industries": {
+    "title": "Индустрии",
+    "subtitle": "Помогаем бизнесу в&nbsp;разработке веб-решений любой сложности",
+    "fintech": {
+      "title": "Финтех",
+      "description": "Обеспечиваем мощное цифровое присутствие для клиентов финансовой сферы."
+    },
+    "eCommerce": {
+      "title": "E-Commerce",
+      "description": "Боремся за&nbsp;каждого клиента в&nbsp;высококонкурентной индустрии интернет-торговли."
+    },
+    "information": {
+      "title": "Информационные технологии",
+      "description": "Выводим IT-компании на&nbsp;новый технологический уровень."
+    },
+    "mediaAndMarketing": {
+      "title": "Медиа и маркетинг",
+      "description": "Помогаем создавать сервисы, к&nbsp;которым хочется возвращаться."
+    }
+  },
   "whatWeDo": {
     "title": "Что мы&nbsp;умеем",
     "list": [
@@ -109,6 +129,10 @@
     "ourAdvantages": [
       ["Разрабатываем по гибким методологиям и берем лучшее от разных подходов."],
       ["Мы сфокусировались на узкой нише разработки фронтенда.", "Нацелены на долгосрочное сотрудничество."]
-    ]
+    ],
+    "fintech": "CSSSR обеспечивает мощное цифровое присутствие для клиентов финансовой сферы.",
+    "eCommerce": "CSSSR борется за каждого клиента в высококонкурентной индустрии интернет-торговли.",
+    "information": "CSSSR — группа компаний, специализирующаяся на фронтенд аутсорсе",
+    "mediaAndMarketing": "CSSSR обеспечивает высокое качество работы с большим объемом данных."
   }
 }
